Update RestAPI_PurchaseTicket
Backend REST API in Node.js for a user purchasing a ticket. This involves generating a unique QR code and sending the ticket back to the user. We will use Express.js for the web server and the QR code library to generate QR codes. This script creates a server that listens for POST requests at the /buy-ticket route. When a request is received, it generates a unique ticket ID, creates a QR code from the ticket info, and sends the QR code image back to the client.